% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generate a map of a state in a specific year of locations where accidents
occurred.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A string that can be used to find the state of interest in
the data frame.}

\item{year}{A string that can be used to find the file of the specific year.}
}
\value{
This function returns NULL if no accidents occurred in the state
during the given year. Otherwise, it returns a scatter plot that highlights the
locations defined by longitude and latitude of accidents occurred.
}
\description{
This function generates a scatter plot map where x and y axes are defined by
longitude and latitude of a state (according to \code{state.num})
in a specific \code{year} to reflect the location of accidents. If the state
cannot be found in the dataset, the function halts and throws an error message.
If there is no accidents in the state for the given year, the function returns
NULL with a message "no accidents to plot".
}
\examples{
\dontrun{fars_map_state("13", "2014")}
\dontrun{fars_map_state("0", "2013")}

}
