% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import data from given filename}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the csv file to be read.}
}
\value{
This function returns a data frame of the data read.
}
\description{
This function reads a csv file defined by \code{filename} if the file
can be found in the directory and returns a data frame. If the file
cannot be found, the function halts and throws "file does not exist" error
message.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}

}
