% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read files from given year and extracts month information in that year for a list
of years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of string of years that will be used to find the files.}
}
\value{
This function returns a list of either data frame where each data frame
contains the MONTH and year column of the dataset or NULL if one or more files
corresponding to year in \code{years} do not exist.
}
\description{
This is a function that reads csv files according to given years and then extracts
month and year information from data. The function throws a warning if accident file for
a year cannot be found.
}
\examples{
fars_read_years({"1996", "2013","2014"})

}
