% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file according to year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string that specifies the year to be included in the filename}
}
\value{
This function returns and prints a string that defines the filename
according to the year that accidents happened
}
\description{
This is a simple function that creates a file according to the \code{year} that
the accidents happened
}
\examples{
\dontrun{make_filename("1996")}
\dontrun{make_filename("2014")}

}
